+++ src/pages/themeproductspage/ThemeProductsList.tsx
@@ -1,81 +1,14 @@
 import styled from "@emotion/styled";
 import { useParams, useNavigate } from "react-router-dom";
-import { useSuspenseApiQuery } from "@/hooks/useSuspenseApiQuery";
-import { useApiQuery } from "@/hooks/useApiQuery";
-import type { Product } from "@/types/api_types";
 import LoadingSpinner from "@/components/common/LoadingSpinner";
-import { useEffect, useRef, useState, useCallback } from "react";
-import useInfiniteScroll from "@/hooks/useInfiniteScroll";
-
-type ThemeProductResponse = {
-  list: Product[];
-  cursor?: string | null;
-  hasMoreList?: boolean;
-};
+import { useThemeProductList } from "./hooks/useThemeProductList";
 
 export default function ThemeProductsList() {
   const { themeId } = useParams<{ themeId: string }>();
   const navigate = useNavigate();
 
-  const [products, setProducts] = useState<Product[]>([]);
-  const [cursor, setCursor] = useState<string | null | undefined>(undefined);
-  const [hasMore, setHasMore] = useState<boolean>(true);
-  const [initLoading, setInitLoading] = useState<boolean>(true);
-  const observerRef = useRef<HTMLDivElement | null>(null);
-
-  const { data: initialData } = useSuspenseApiQuery<ThemeProductResponse>({
-    url: `/api/themes/${themeId}/products`,
-    queryKey: ["theme-products", themeId, "initial"],
-  });
-
-  const { data, isLoading, isError, error, refetch } =
-    useApiQuery<ThemeProductResponse>({
-      url: `/api/themes/${themeId}/products${cursor ? `?cursor=${cursor}` : ""}`,
-      queryKey: ["theme-products", themeId, cursor],
-      enabled: false,
-    });
-
-  useEffect(() => {
-    if (initialData) {
-      setProducts(initialData.list);
-      setCursor(initialData.cursor ?? null);
-      setHasMore(
-        initialData.hasMoreList !== false && !!initialData.list.length
-      );
-      setInitLoading(false);
-    }
-  }, [initialData]);
-
-  useEffect(() => {
-    if (isError && error) {
-      throw error;
-    }
-  }, [isError, error]);
-
-  useEffect(() => {
-    if (data && cursor) {
-      setProducts((prev) => {
-        const existingIds = new Set(prev.map((item) => item.id));
-        const filtered = data.list.filter((item) => !existingIds.has(item.id));
-        return [...prev, ...filtered];
-      });
-      setCursor(data.cursor ?? null);
-      setHasMore(data.hasMoreList !== false && !!data.list.length);
-    }
-  }, [data]);
-
-  const handleObserver = useCallback(() => {
-    if (hasMore && !isLoading && !initLoading && products.length > 0) {
-      refetch();
-    }
-  }, [hasMore, isLoading, initLoading, products.length, refetch]);
-
-  useInfiniteScroll({
-    targetRef: observerRef,
-    onIntersect: handleObserver,
-    enabled: hasMore && !isLoading && !initLoading && products.length > 0,
-    threshold: 0.4,
-  });
+  const { products, isLoading, initLoading, hasMore, observerRef } =
+    useThemeProductList({ themeId });
 
   const handleItemClick = (id: number) => {
     navigate(`/order/${id}`);
